Enhance global styles & component hierarchy
