+++ frontend/styles/globals.css
@@ -77,7 +77,7 @@
   }
 
   * {
-    @apply border-border;
+    @apply border-border outline-ring/50 font-base;
   }
 
   html {
@@ -85,19 +85,12 @@
   }
 
   body {
-    @apply w-full font-base text-foreground antialiased;
+    @apply bg-background text-foreground w-full antialiased;
   }
 
   .wavy-background {
     background-image: url('/background.avif');
     background-repeat: repeat;
     background-size: 100% auto;
-  }
-
-  * {
-    @apply border-border outline-ring/50;
-  }
-  body {
-    @apply bg-background text-foreground;
   }
 }
